+++ CHANGELOG.md
@@ -29,6 +29,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `IdentificationInfo` resource and `EnvelopeRecipient.identification`
 - Add `EnvelopeProperties.auditLogAvailableToAccountUsers`
 - Add `IdentificationInfo.approvalMode`
+- Add `EnvelopeRecipient.approvalMode`
+- Add `EnvelopeRecipient.approveDocumentsAtOnce`
+- Add `EnvelopeTemplateRecipient.approvalMode`
+- Add `EnvelopeTemplateRecipient.approveDocumentsAtOnce`
 
 ### Changed
 - Change `Identification.envelope` from string to EnvelopeInfo