+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `Identification.approve` endpoint and `Identification.deny` endpoint
 - Add `notificationForSenderEmail` to `AccountSettings`
 - Add `EnvelopeTag.format`, `EnvelopeTag.dateOfSignature`, `EnvelopeTemplateTag.format`
+- Add `ScenarioVersionInfo` to `Identification`
 
 ### Changed
 - Change `EnvelopeDocument.file` to nullable