+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `MyContactsEndpoint.list` endpoint
 - Add `MyContactsEndpoint.update` endpoint
 - Add `MyContactsEndpoint.delete` endpoint
+- Add `IdentifyScenarioVersion.ownConditions`
 - Add `MyAccount.shortName`
 - Add `MyEndpoint.info` endpoint and `MyInfo` resource
 - Add `MyAccountEndpoint.switch` endpoint