+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `IdentifyScenarioVersion.selfieEnabled`
+- Add `AccountBilling.identifyAi`
 - Add `MyAccount.createdAt`
 - Add `MyContactsEndpoint` CRUD endpoints
 - Add `IdentifyScenarioVersion.ownConditions`